Made Kudos objects
Instead of an array that only contains a kudos message, now they act as objects with an associated author, message, and source.  This is currently represented in a simplistic fashion, so should be changed in the future so that the message, author and source are treated as separate elements.  This would mean that different styles could be applied to them.