Redesigned layout
"About Us" moved to main, and main content has been reformatted so that the image and text box are inline with each other.
Moved sidebar to bottom of main content, will be a horizontal band for the future with better styling and formatting for varying screen sizes.  This main content is contained in the "splash" class.