Updated form + misc changes
Using XAMPP, a form was successfully able to send its data to a locally hosted server using php. Currently, resources.html contains the form, then sends the data to form_submit.php. In that php file is HTML which confirms the form was successfully sent including elements present on other pages (navbar, footer). Currently, the navbar does not work properly because the links are not set up properly.