Form and Website Styling Changes
Added a section of code which should generate email messages based off of the user's form input. Currently it is not functional due to an issue with how the mail server is configured. Once that issue is resolved, it should be able to send an email to the employer's defined email in addition to a confirmation email to the user.
Work has begun on changing how the website's styling changes for different environment (e.g. phones). It is still a work-in-progress, and currently the website still appears strangely on mobile resolutions.