Fixed scaling issues for mobile view
The website is much more responsive and is better suited for mobile resolutions now.  Additionally, the navbar is now 'sticky' for the mobile version, so that issue has been resolved. The website should work across a greater range of resolutions and aspect ratios as a result.
As well, 'kudos' has been removed because it will not be implemented in the first release of this website, and will be added later once a database that can store all the relevant data can be set up. In addition, a UI for them to be added and removed will also have to be implemented at that time so that a user can sign in and change the database from the website itself.