Minor style changes
All main headings should be the same style now. Form styling was changed to look more appealing and be more responsive (e.g. input field changes color when selected). As well, images should display 100% of their available area now, although this may need to be adjusted in the future so that images are not too large at larger resolutions.