Improved 'Services' section
Services are now displayed as gridlike elements, all of the same width and size. On the standard desktop view, there are 2 columns, whereas the mobile displays only 1 column for legibility.  A consideration at the moment is for 3 columns to be displayed on resolutions greater than 1080p, but that has not yet been implemented or tested.
Images now no longer stretch beyond 1000px, meaning that for larger displays, it should no longer overwhelm the webpage.
Work has begun on properly implementing a file upload system in PHP. At the moment, it can only successfully determine if a file has been uploaded, however all of the supplemental code which is intended to validate it (prevent files that are too large or not one of the accepted file types) does not functional at the moment.