Misc changes
Added descriptions to services section to make it more clear what services are provided.
Improved the input field selection for form_feedback.  Form fields are saved into a new object defined by the 'Field' class which currently includes: the name of the form input, the HTML element itself, the maximum characters of the input field, and a method for determining whether or not the input field is empty.
ISSUE: Submitting a form opens up a new tab with an empty page, this should change in the future so that submitting a form with invalid inputs should display warning messages to the user and prevent running the PHP script.