Improved client-side form validation
Phone numbers are now validated on the client side. As well, HTML now specifies what files are accepted when a user uploads a file, and javascript will verify that the file has the correct file extension before allowing the form to post. However, this does not prevent someone from uploading 'dummy' files which simply have the file extension changed by the user (e.g. a .txt file being changed to a .pdf is still accepted).