Added server-side validation
First and last name are now validated client-side to only contain alphabetical characters. As well, the phone number is verified to only contain numbers.
Cleaned up php and javascript code by removing unnecessary variables or switch cases (namely the parts of the code that dealt with validating comments because the comment section will not be validated as of this moment).